# swagger.yaml
openapi: 3.0.0
info:
  title: kubuka kwesu tobecisi cabutonga
  description: Kubuka is an organisation
  version: 1.0.0
servers:
  - url: https://server.kubukaspace.com
    description: Online development server
paths:
...
  /contactus:
    post:
      summary: Leave a message for admins to contact you latter
      description: Save messages from visitor
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                names:
                  type: string
                  example: "Mwana wa Kubuka"
                email:
                  type: string
                  example: "info@kubuka.com"
                message:
                  type: string
                  example: "Thanks kubuka, I would like to invest."
              required:
                - names
                - email
                - message
      responses:
        "200":
          description: The email has been successfully subscribed to the Basilwizi and will receive regular updates on news and latest stories
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message was went to Admin"
    get:
      summary: Get a list of all messages
      description: Restricted to admin users.
      operationId: getAllContactus
      responses:
        "200":
          description: An array of all messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    names:
                      type: string
                      example: "Mwana wa Buka"
                    email:
                      type: string
                      example: "info@kubuka.com"
                    message:
                      type: string
                      example: "Thanks kubuka, I would like to invest."
                    created:
                      type: string
                      example: "2020-05-05T09:12:57.848Z"
                    updated:
                      type: string
                      example: "2020-05-08T03:11:21.553Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete a Message
      description: Admin users can delete any Message, regular users are restricted to their own Message.
      operationId: deleteCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Message deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"